Don't have the Helioselene Ciphersuite use the provided read_G
The provided read_G compares the point's serialization against its
reserialization to check if the encoding was malleated. helioselene already
checks the point's encoded x coordinate is reduced and the point's sign bit
isn't for the representation of identity, preventing malleation. We accordingly
don't need this additional check which is expensive.